changing state from AppBar Drawer working - now to refresh dataVis
From the MenuItem's in the AppBar's Drawer, we can now define what the current dataset is and the title of the page by passing through the information the button is pertaining to.

An optimal way to go about this would be to generate the buttons by passing throuhg a map or equiv. with the TITLE and ENDPOINT pairs for cleaner execution. I'll save that for a rainy day (later date)